pyproject.toml
[project]
name = "nanochat"
version = "0.1.0"
description = "the minimal full-stack ChatGPT clone"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.0.0",
    "fastapi>=0.117.1",
    "files-to-prompt>=0.6",
    "numpy==1.26.4",
    "psutil>=7.1.0",
    "regex>=2025.9.1",
    "setuptools>=80.9.0",
    "torch>=2.8.0",
    "uvicorn>=0.36.0",
    "wandb>=0.21.3",
    "Pillow>=10.0.0",
    "torchvision>=0.18.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# target torch to cuda 12.8 or CPU
[tool.uv.sources]
torch = [  
    { index = "pytorch-cpu", extra = "cpu" },  
    { index = "pytorch-cu128", extra = "gpu" },  
]

[[tool.uv.index]]  
name = "pytorch-cpu"  
url = "https://download.pytorch.org/whl/cpu"  
explicit = true  
  
[[tool.uv.index]]  
name = "pytorch-cu128"  
url = "https://download.pytorch.org/whl/cu128"  
explicit = true

[project.optional-dependencies]  
cpu = [  
    "torch>=2.8.0",  
]  
gpu = [  
    "torch>=2.8.0",  
]  
  
[tool.uv]  
conflicts = [  
    [  
        { extra = "cpu" },  
        { extra = "gpu" },  
    ],  
]  